move env variables to envConfig.js
