+++ microservices/strapi-connector-ms/app.js
@@ -4,62 +4,7 @@
 const fastify = require('fastify')
 const fastifyEnv = require('@fastify/env')
 
-const envSchema = {
-    type: 'object',
-    required: ['SERVER_SERVLET_CONTEXT_PATH',
-                'SPRING_DATASOURCE_USERNAME',
-                'SPRING_DATASOURCE_PASSWORD',
-                'SPRING_DATASOURCE_URL',
-                'APP_ORIGIN',
-                'KEYCLOAK_REALM',
-                'KEYCLOAK_AUTH_URL',
-                'KEYCLOAK_CLIENT_SECRET',
-                'KEYCLOAK_CLIENT_ID'
-            ],
-    properties: {
-        SERVER_SERVLET_CONTEXT_PATH: {
-            type: 'string',
-            default: '/'
-        },
-        SPRING_DATASOURCE_USERNAME: {
-            type: 'string',
-            default: 'entando-strapi-connector-user'
-        },
-        SPRING_DATASOURCE_PASSWORD: {
-            type: 'string',
-            default: 'entando-strapi-connector-password'
-        },
-        SPRING_DATASOURCE_URL: {
-            type: 'string',
-            default: 'jdbc:postgresql://localhost:5432/entando-strapi-connector'
-        },
-        APP_ORIGIN: {
-            type: 'string',
-            default: 'http://localhost:3000'
-        },
-        KEYCLOAK_REALM: {
-            type: 'string',
-            default: 'entando-dev'
-        },
-        KEYCLOAK_AUTH_URL: {
-            type: 'string',
-            default: 'http://localhost:9080/auth'
-        },
-        KEYCLOAK_CLIENT_SECRET: {
-            type: 'string',
-            default: 'CFXd3zdoDgnSvLfovOvjFzpFkfnWfFar'
-        },
-        KEYCLOAK_CLIENT_ID: {
-            type: 'string',
-            default: 'entando-strapi-connector-ms-client'
-        }
-    }
-}
-
-const envOptions = {
-    confKey: 'config', // optional, default: 'config'
-    schema: envSchema,
-}
+const envOptions = require('./config/envConfig')
 
 function build(opts = {}) {
     const app = fastify(opts)
@@ -79,7 +24,7 @@ function build(opts = {}) {
             }
             app.register(require('./plugins/postgres'), app.config)
             app.register(require('./routes/api/health'), healthOpts)
-            app.register(require('./routes/api/strapi-config'), strapiConfigOpts)
+            app.register(require('./routes/api/strapiConfig'), strapiConfigOpts)
         }
     )
